Fix bug where old tiles are not removed
Some tiles would linger on the screen after the level was changed. This
was due to the fact the render struct caches an image to reuse if the
player hasn't moved (no world events). Instead of clearing the screen on
redraw, the render would just write over what was there. This doesn't
matter when the tiles don't change, but when we need to change levels,
the image needs to be completely redrawn. To prevent future issues, the
render struct now clears the cached image on every redraw.